backend/requirements.txt
Django==4.1.1
daphne==4.0.0
channels==4.0.0
attrs==22.2.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6